Take 1.6 out of the build matrix
Go 1.6 doesn't come with a test case runner which I've started using in
the project. I might be able to backpatch it, but the easiest way to go
forward is to take 1.6 out. Go versions tend to be very backwards
compatible, so this is quite unlikely to cause any significant problems.