Add strict mode for form encoding
I'm pretty sure that we've checked every possible params struct for
malformed form tags, but this modifies the encoding scheme so that we
only ever panic when strict mode is engaged. Strict mode is disabled by
default, but we enable it throughout the entire test suite to maximize
our visibility into problems.

In case we did miss a param struct that's in use by a real user, at
least their program won't panic after they upgrade.